backend/requirements.txt
djangorestframework==3.15.2
python-dotenv==1.0.1
django-cors-headers==4.4.0
psycopg2==2.9.9
